+++ java-generator/core/src/main/java/io/fabric8/java/generator/nodes/JObject.java
@@ -48,7 +48,7 @@ public class JObject extends AbstractJSONSchema2Pojo implements JObjectExtraAnno
   protected final String className;
   protected final String pkg;
   protected final Map<String, AbstractJSONSchema2Pojo> fields;
-  private final Set<String> required;
+  protected final Set<String> required;
   private final Set<String> deprecated = new HashSet<>();
 
   private final boolean preserveUnknownFields;