+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 #### Dependency Upgrade
 
 #### New Features
+* Fix #6802: Java generator support for required spec and status
 
 * Fix #5993: Support for Kubernetes v1.31 (elli)
 * Fix #6767: Support for Kubernetes v1.32 (penelope)